Factor out phase creation from test for docs
